Normalize joystick values passed to Unity

+++ Android/Roll-a-Ball/src/main/java/com/github/artnest/rollaball/PlayerMessageHandler.kt
@@ -1,5 +1,5 @@
 package com.github.artnest.rollaball
 
 interface PlayerMessageHandler {
-    fun onMoved(x: Float, y: Float)
+    fun onMoved(x: Int, y: Int)
 }
